docs(habitat): consolidate G4 recipe import matrix
Replace the overlapping G4 recipe domain import proxy packets with one positive Habitat Grit rule for standard recipe source. The surviving rule keeps import-surface authority in Habitat/Grit, deletes the absorbed proxy packets, repairs the current morphology constant import through the public ops surface, and updates the canonical remediation JSON plus Domino 58 receipt.\n\nVerification:\n- bun habitat check --rule require_public_domain_surfaces_in_recipes_and_maps --json\n- bun run --cwd tools/habitat test test/lib/pattern-apply.test.ts test/rules/pattern-views.test.ts test/service/fix-service.test.ts test/rules/registry/contract.test.ts\n- nx run mod-swooper-maps:check\n- bun run --cwd tools/habitat check\n- bun habitat classify .habitat\n- live manifest/support/runner path resolution against canonical JSON\n- git diff --check
