--- .habitat/blueprints/domain/prohibit_relative_domain_reaches_from_recipes_and_maps/baseline.json
@@ -1,1 +1,0 @@
-[]
