--- .habitat/blueprints/domain/prohibit_relative_domain_reaches_from_recipes_and_maps/rule.json
@@ -1,42 +1,0 @@
-{
-  "schemaVersion": 2,
-  "id": "prohibit_relative_domain_reaches_from_recipes_and_maps",
-  "title": "Prohibit Relative Domain Reaches From Recipes And Maps",
-  "placement": {
-    "niche": "civ7/mapgen/domains",
-    "blueprint": "domain",
-    "category": "boundary"
-  },
-  "operation": {
-    "kind": "check"
-  },
-  "ownerProject": "mod-swooper-maps",
-  "lane": "enforced",
-  "forbids": "recipe/map source importing local domain source through relative ../domain paths",
-  "why": "Recipes and maps compose domains through named public surfaces, not local source-tree reaches.",
-  "remediate": null,
-  "message": "Import domain code through @mapgen/domain/<domain>, /ops, or /config.js rather than relative ../domain paths.",
-  "pathCoverage": [
-    {
-      "kind": "exact-path",
-      "patterns": [
-        "mods/*/src/recipes/**/*.ts",
-        "mods/*/src/maps/**/*.ts",
-        "mods/*/src/recipes/**/*.tsx",
-        "mods/*/src/maps/**/*.tsx"
-      ]
-    }
-  ],
-  "scanRoots": ["mods/mod-swooper-maps/src/recipes", "mods/mod-swooper-maps/src/maps"],
-  "hookCheck": true,
-  "supportFiles": {
-    "baseline": ".habitat/blueprints/domain/prohibit_relative_domain_reaches_from_recipes_and_maps/baseline.json"
-  },
-  "runner": {
-    "name": "grit",
-    "files": {
-      "pattern": ".habitat/blueprints/domain/prohibit_relative_domain_reaches_from_recipes_and_maps/pattern.md"
-    },
-    "patternName": "prohibit_relative_domain_reaches_from_recipes_and_maps"
-  }
-}
