--- .habitat/blueprints/domain/restrict_recipes_to_public_domain_surfaces/baseline.json
@@ -1,1 +1,0 @@
-[]
