--- .habitat/blueprints/domain/restrict_recipes_to_public_domain_surfaces/rule.json
@@ -1,37 +1,0 @@
-{
-  "schemaVersion": 2,
-  "id": "restrict_recipes_to_public_domain_surfaces",
-  "title": "Restrict Recipes To Public Domain Surfaces",
-  "placement": {
-    "niche": "civ7/mapgen/domains",
-    "blueprint": "domain",
-    "category": "boundary"
-  },
-  "operation": {
-    "kind": "check"
-  },
-  "ownerProject": "mod-swooper-maps",
-  "lane": "enforced",
-  "forbids": "recipes importing non-public domain subpaths",
-  "why": "Recipes compose domains only through stable public surfaces.",
-  "remediate": null,
-  "message": "Recipes may import @mapgen/domain/<domain>, @mapgen/domain/<domain>/ops, or /config.js only.",
-  "pathCoverage": [
-    {
-      "kind": "exact-path",
-      "patterns": ["mods/mod-swooper-maps/src/recipes/**/*.ts"]
-    }
-  ],
-  "scanRoots": ["mods/mod-swooper-maps/src/recipes"],
-  "hookCheck": true,
-  "supportFiles": {
-    "baseline": ".habitat/blueprints/domain/restrict_recipes_to_public_domain_surfaces/baseline.json"
-  },
-  "runner": {
-    "name": "grit",
-    "files": {
-      "pattern": ".habitat/blueprints/domain/restrict_recipes_to_public_domain_surfaces/pattern.md"
-    },
-    "patternName": "restrict_recipes_to_public_domain_surfaces"
-  }
-}
