+++ docs/system/libs/mapgen/policies/IMPORTS.md
@@ -99,7 +99,5 @@ This policy is the simplest guardrail that keeps the ecosystem coherent: use the
 - Exported entrypoints (source of truth for allowed imports): `packages/mapgen-core/package.json`
 - `@mapgen/*` is an internal/workspace alias, used inside the package: `packages/mapgen-core/src/engine/index.ts`
 - Target posture for packaging and boundaries: `docs/projects/engine-refactor-v1/resources/spec/SPEC-packaging-and-file-structure.md`
-- Recipe import guard: Habitat `grit-recipe-domain-surface` and
-  `grit-domain-deep-import` rules in
-  `.habitat/blueprints/domain/restrict_recipes_to_public_domain_surfaces/rule.json` and
-  `.habitat/blueprints/domain/prohibit_relative_domain_reaches_from_recipes_and_maps/rule.json`
+- Recipe import guard: Habitat Grit rule in
+  `.habitat/blueprints/domain/require_public_domain_surfaces_in_recipes_and_maps/rule.json`
