+++ mods/mod-swooper-maps/src/domain/morphology/ops/index.ts
@@ -45,6 +45,7 @@ const implementations = {
 export default implementations;
 
 export type { MountainsConfig } from "./mountains-shared/config.js";
+export { DEFAULT_ELEVATION_SCALE } from "./compute-base-topography/rules/index.js";
 export {
   computeBaseTopography,
   computeBeltDrivers,
