+++ tools/habitat/test/rules/registry/contract.test.ts
@@ -31,14 +31,14 @@ describe("rule registry contract", () => {
       readText: readTextSync,
     }).rules;
 
-    expect(rules).toHaveLength(126);
-    expect(rules.filter((rule) => rule.runner.name === "grit")).toHaveLength(79);
+    expect(rules).toHaveLength(118);
+    expect(rules.filter((rule) => rule.runner.name === "grit")).toHaveLength(71);
     expect(
       rules.filter((rule) => rule.runner.name === "habitat" && rule.runner.mode === "script")
-    ).toHaveLength(33);
+    ).toHaveLength(34);
     expect(
       rules.filter((rule) => rule.runner.name === "habitat" && rule.runner.mode === "structure")
-    ).toHaveLength(8);
+    ).toHaveLength(7);
     expect(
       rules.filter((rule) => rule.runner.name === "habitat" && rule.runner.mode === "file-layer")
     ).toHaveLength(5);
